rfactor: implements ClueManager
