+++ src/main/java/managers/RoomManager.java
@@ -6,14 +6,12 @@ import com.mongodb.client.MongoCursor;
 import com.mongodb.client.model.Filters;
 import com.mongodb.client.model.Projections;
 import database.MongoDBConnection;
-import dtos.ClueDTO;
 import dtos.DecorationDTO;
 import dtos.RoomDTO;
-import entities.Clue;
 import entities.Decoration;
 import entities.Room;
 import entities.enums.Difficulty;
-import entities.enums.Theme;
+import lombok.Getter;
 import org.bson.Document;
 import org.bson.types.ObjectId;
 import utils.ValidateInputs;
@@ -21,8 +19,9 @@ import utils.ValidateInputs;
 import java.util.ArrayList;
 import java.util.List;
 
+@Getter
 public class RoomManager {
-    MongoCollection<Document> escapeRoomCollection = MongoDBConnection.getEscapeRoomCollection();
+    private final MongoCollection<Document> escapeRoomCollection = MongoDBConnection.getEscapeRoomCollection();
 
     public static Room createRoom() {
 
@@ -154,67 +153,6 @@ public class RoomManager {
         }
     }
 
-    public List<ClueDTO> getAllCluesDTO(ObjectId roomId) {
-        List<ClueDTO> clueDTOS = new ArrayList<>();
-        try {
-            Document roomDoc = escapeRoomCollection.find(Filters.eq("_id", roomId))
-                    .projection(Projections.include("clues"))
-                    .first();
-
-            if (roomDoc != null && roomDoc.containsKey("clues")) {
-                List<Document> clues = roomDoc.getList("clues", Document.class);
-                for (Document clue : clues) {
-                    clueDTOS.add(new ClueDTO(
-                            clue.getObjectId("_id"),
-                            clue.getString("name")
-                    ));
-                }
-            }
-        } catch (MongoException e) {
-            System.out.println("Database error: " + e.getMessage());
-        }
-        return clueDTOS;
-    }
-
-    public int ChosenDTOClue(String action, ObjectId roomId) {
-        List<ClueDTO> clues = getAllCluesDTO(roomId);
-        if (clues.isEmpty()) {
-            System.out.println("No clues to " + action + " in this room.");
-            return 0;
-        } else {
-            for (int i = 0; i < clues.size(); i++) {
-                System.out.println((i + 1) + ". " + clues.get(i).getName());
-            }
-            System.out.println("0. Go back");
-            return ValidateInputs.validateIntegerBetweenOnRange(
-                    "Choose the clue you want to " + action + ": ", 0, clues.size());
-        }
-    }
-
-    public void deleteClueFromRoom() {
-        List<RoomDTO> rooms = getAllRoomsDTO();
-        int roomChoice = ChosenDTORoom("delete clue");
-        if (roomChoice == 0) {
-            System.out.println("Going back...");
-        } else {
-            RoomDTO room = rooms.get(roomChoice - 1);
-            ObjectId roomId = room.getId();
-
-            int clueChoice = ChosenDTOClue("delete", roomId);
-            if (clueChoice == 0) {
-                System.out.println("Going back...");
-            } else {
-                List<ClueDTO> clues = getAllCluesDTO(roomId);
-                ClueDTO clue = clues.get(clueChoice - 1);
-
-                escapeRoomCollection.updateOne(
-                        Filters.eq("_id", roomId),
-                        new Document("$pull", new Document("clues", new Document("_id", clue.getId())))
-                );
-                System.out.println(">> Clue '" + clue.getName() + "' successfully deleted.");
-            }
-        }
-    }
 
     public List<DecorationDTO> getAllDecorationsDTO(ObjectId roomId) {
         List<DecorationDTO> decorationsDAO = new ArrayList<>();
@@ -278,38 +216,100 @@ public class RoomManager {
         }
     }
 
-    public static Clue createClue() {
-        String name = ValidateInputs.validateString("Enter the name of the clue to create: ");
-        int price = ValidateInputs.validateIntegerBetweenOnRange(
-                "Enter the clue price in €: ", 1, 999999
-        );
-        Theme theme = ValidateInputs.validateEnum(Theme.class, "Enter theme (PUZZLE, PASSWORD, SYMBOL): ");
-        return Clue.builder()
-                .id(new ObjectId())
-                .price(price)
-                .name(name)
-                .theme(theme)
-                .build();
-    }
+//    public static Clue createClue() {
+//        String name = ValidateInputs.validateString("Enter the name of the clue to create: ");
+//        int price = ValidateInputs.validateIntegerBetweenOnRange(
+//                "Enter the clue price in €: ", 1, 999999
+//        );
+//        Theme theme = ValidateInputs.validateEnum(Theme.class, "Enter theme (PUZZLE, PASSWORD, SYMBOL): ");
+//        return Clue.builder()
+//                .id(new ObjectId())
+//                .price(price)
+//                .name(name)
+//                .theme(theme)
+//                .build();
+//    }
+//
+//    public void addClueToRoom() {
+//        List<RoomDTO> rooms = getAllRoomsDTO();
+//        int roomChoice = ChosenDTORoom("add clue");
+//        if (roomChoice == 0) {
+//            System.out.println("Going back...");
+//        } else {
+//            RoomDTO room = rooms.get(roomChoice - 1);
+//            Clue clue = createClue();
+//            Document clueDoc = new Document("_id", clue.getId())
+//                    .append("price", clue.getPrice())
+//                    .append("name", clue.getName())
+//                    .append("theme", clue.getTheme().toString());
+//            escapeRoomCollection.updateOne(
+//                    Filters.eq("_id", room.getId()),
+//                    new Document("$push", new Document("clues", clueDoc))
+//            );
+//            System.out.println(">> Clue '" + clue.getName() + "' added to room '" + room.getName() + "'");
+//        }
+//    }
 
-    public void addClueToRoom() {
-        List<RoomDTO> rooms = getAllRoomsDTO();
-        int roomChoice = ChosenDTORoom("add clue");
-        if (roomChoice == 0) {
-            System.out.println("Going back...");
-        } else {
-            RoomDTO room = rooms.get(roomChoice - 1);
-            Clue clue = createClue();
-            Document clueDoc = new Document("_id", clue.getId())
-                    .append("price", clue.getPrice())
-                    .append("name", clue.getName())
-                    .append("theme", clue.getTheme().toString());
-            escapeRoomCollection.updateOne(
-                    Filters.eq("_id", room.getId()),
-                    new Document("$push", new Document("clues", clueDoc))
-            );
-            System.out.println(">> Clue '" + clue.getName() + "' added to room '" + room.getName() + "'");
-        }
-    }
+//    public List<ClueDTO> getAllCluesDTO(ObjectId roomId) {
+//        List<ClueDTO> clueDTOS = new ArrayList<>();
+//        try {
+//            Document roomDoc = escapeRoomCollection.find(Filters.eq("_id", roomId))
+//                    .projection(Projections.include("clues"))
+//                    .first();
+//
+//            if (roomDoc != null && roomDoc.containsKey("clues")) {
+//                List<Document> clues = roomDoc.getList("clues", Document.class);
+//                for (Document clue : clues) {
+//                    clueDTOS.add(new ClueDTO(
+//                            clue.getObjectId("_id"),
+//                            clue.getString("name")
+//                    ));
+//                }
+//            }
+//        } catch (MongoException e) {
+//            System.out.println("Database error: " + e.getMessage());
+//        }
+//        return clueDTOS;
+//    }
+//
+//    public int ChosenDTOClue(String action, ObjectId roomId) {
+//        List<ClueDTO> clues = getAllCluesDTO(roomId);
+//        if (clues.isEmpty()) {
+//            System.out.println("No clues to " + action + " in this room.");
+//            return 0;
+//        } else {
+//            for (int i = 0; i < clues.size(); i++) {
+//                System.out.println((i + 1) + ". " + clues.get(i).getName());
+//            }
+//            System.out.println("0. Go back");
+//            return ValidateInputs.validateIntegerBetweenOnRange(
+//                    "Choose the clue you want to " + action + ": ", 0, clues.size());
+//        }
+//    }
 
+//    public void deleteClueFromRoom() {
+//        List<RoomDTO> rooms = getAllRoomsDTO();
+//        int roomChoice = ChosenDTORoom("delete clue");
+//        if (roomChoice == 0) {
+//            System.out.println("Going back...");
+//        } else {
+//            RoomDTO room = rooms.get(roomChoice - 1);
+//            ObjectId roomId = room.getId();
+//
+//            int clueChoice = ChosenDTOClue("delete", roomId);
+//            if (clueChoice == 0) {
+//                System.out.println("Going back...");
+//            } else {
+//                List<ClueDTO> clues = getAllCluesDTO(roomId);
+//                ClueDTO clue = clues.get(clueChoice - 1);
+//
+//                escapeRoomCollection.updateOne(
+//                        Filters.eq("_id", roomId),
+//                        new Document("$pull", new Document("clues", new Document("_id", clue.getId())))
+//                );
+//                System.out.println(">> Clue '" + clue.getName() + "' successfully deleted.");
+//            }
+//        }
+//    }
 }
+
