+++ src/main/java/utils/Menus.java
@@ -4,10 +4,7 @@ import daos.RoomDAOImpl;
 import database.MongoDBConnection;
 import entities.Player;
 import entities.Room;
-import managers.BusinessManager;
-import managers.CertificatePrinter;
-import managers.EscapeRoomManager;
-import managers.RoomManager;
+import managers.*;
 
 public class Menus {
     // Menu provisional
@@ -46,7 +43,7 @@ public class Menus {
                 """, 0, 4);
     }
 
-    public static void mainMenuManager(EscapeRoomManager escapeRoomManager, RoomManager roomManager, RoomDAOImpl roomDAOImpl, BusinessManager businessManager) {
+    public static void mainMenuManager(EscapeRoomManager escapeRoomManager, RoomManager roomManager, RoomDAOImpl roomDAOImpl, BusinessManager businessManager,ClueManager clueManager) {
         boolean exit = false;
 
         while (!exit) {
@@ -70,7 +67,7 @@ public class Menus {
                     roomManager.deleteRoomByUserSelection();
                     break;
                 case 5:
-                    subMenuEditManager(roomManager);
+                    subMenuEditManager(roomManager, clueManager);
                     break;
                 case 6:
                     // TODO: decidir flujo de creación jugador/venta ticket ¿?
@@ -104,15 +101,15 @@ public class Menus {
         }
     }
 
-    public static void subMenuEditManager(RoomManager roomManager) {
+    public static void subMenuEditManager(RoomManager roomManager, ClueManager clueManager) {
         boolean exit = false;
         do {
             switch (subMenuEditOptions()) {
                 case 1:
-                    roomManager.addClueToRoom();
+                    clueManager.addClueToRoom();
                     break;
                 case 2:
-                    roomManager.deleteClueFromRoom();
+                    clueManager.deleteClueFromRoom();
                     break;
                 case 3:
                     roomManager.addDecorationToRoom();
